+++ Src/Metrics.Tests/Visualization/MetricsHttpListenerTests.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using System.Linq;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
@@ -17,7 +17,7 @@ namespace Metrics.Tests.Visualization
         private static Task<MetricsHttpListener> StartListener(string endpoint)
         {
             var context = new TestContext();
-            return MetricsHttpListener.StartHttpListenerAsync(Endpoint(endpoint), context.DataProvider, () => new HealthStatus(), CancellationToken.None);
+            return MetricsHttpListener.StartHttpListenerAsync(Endpoint(endpoint), Enumerable.Empty<MetricsEndpoint>(), CancellationToken.None);
         }
 
         [Fact]
@@ -26,7 +26,7 @@ namespace Metrics.Tests.Visualization
             var listener = await StartListener("HttpEndpointCanBeDisposed");
             listener.Dispose();
         }
-        
+
         [Fact]
         public async Task MetricsHttpListener_CanBeDoubleDisposed()
         {
@@ -43,7 +43,7 @@ namespace Metrics.Tests.Visualization
             {
                 listener.Prefixes.Add(Endpoint(endpoint));
                 listener.Start();
-                
+
                 var result = await StartListener(endpoint);
                 result.Should().BeNull();
             }
@@ -59,13 +59,15 @@ namespace Metrics.Tests.Visualization
                 listener.Start();
 
                 var loggedAnError = false;
-                using (var config = Metric.Config)
+
+                using (var config = CreateConfig())
                 {
                     config.WithErrorHandler((exception, s) => { loggedAnError = true; }, true);
                     config.WithErrorHandler((exception) => { loggedAnError = true; }, true);
 
                     await StartListener(endpoint);
                 }
+
                 Assert.True(loggedAnError);
                 listener.Close();
             }
@@ -75,10 +77,16 @@ namespace Metrics.Tests.Visualization
         [Fact]
         public void MetricsHttpListener_MetricsConfig_SecondCallToWithHttpEndportDoesNotThrow()
         {
-            using (var config = Metric.Config.WithHttpEndpoint("http://localhost:58888/metricstest/HttpListenerTests/sameendpoint/"))
+            using (var config = CreateConfig().WithHttpEndpoint("http://localhost:58888/metricstest/HttpListenerTests/sameendpoint/"))
             {
                 config.WithHttpEndpoint("http://localhost:58888/metricstest/HttpListenerTests/sameendpoint/");
             }
+        }
+
+        private static MetricsConfig CreateConfig()
+        {
+            var context = new DefaultMetricsContext("http-listener-tests");
+            return new MetricsConfig(context);
         }
     }
 }