+++ Src/Metrics/ElasticSearch/ElasticSearchConfigExtensions.cs
@@ -6,10 +6,18 @@ namespace Metrics
 {
     public static class ElasticSearchConfigExtensions
     {
+
+        public static MetricsReports WithElasticSearch(this MetricsReports reports, ElasticReportsConfig reportConfig, TimeSpan interval)
+        {
+            return reports.WithReport(new ElasticSearchReport(reportConfig), interval);
+        }
+
+        [Obsolete("Use WithElasticSearch(ElasticReportsConfig) instead")]
         public static MetricsReports WithElasticSearch(this MetricsReports reports, string host, int port, string index, TimeSpan interval)
         {
-            var uri = new Uri($@"http://{host}:{port}/_bulk");
-            return reports.WithReport(new ElasticSearchReport(uri, index), interval);
+            return reports.WithElasticSearch(new ElasticReportsConfig() { Index = index, Host = host, RollingIndexType = RollingIndexType.None, Port = port }, interval);
         }
+
+
     }
 }